requirements.txt
geopandas
rasterio
langchain
langchain-community
langchain-core
streamlit
langchain-google-genai
python-dotenv
matplotlib
python-dotenv
folium 
osmnx 
geopy 
networkx
scikit-learn